pizza_factory_locator v0.1.0 Database

The mongodb Database API module. All communication to the database should be done through this module.

Functions

connect()

Attempts to connect to the Mongo Database using the configuration file

get_factories(start_index, stop_index, zone_coordinates_start \\ nil, zone_coordinates_stop \\ nil)

Gets a slice of factories from the database. The size of the slice and the region to fetch factories from can be specified as parameters

get_factory_count()

Gets the number of factories from the database

get_order_count()

Gets the number of orders from the database

get_orders(start_index, stop_index, zone_coordinates_start \\ nil, zone_coordinates_stop \\ nil)

Fetches orders from the database. Returns an array of Order structs. Orders from a particular area may be fetched by passing zone parameters. Start coordinates indicate the top left of a geographical area while end coordinates indicate the bottom right.

save_factory(factory)

Saves a factory to the database

save_order(order)

Saves an order to the database