pizza_factory_locator v0.1.0 PizzaFactoryLocator

Module containing the main functionality of the project, that is the "determine_new_factory_location" function. This function should be exposed via a command-line API at deployment so that it can be intergrated into other projects

Functions

determine_new_factory_location(boundary_start \\ nil, boundary_stop \\ nil)

Parses large order dataset stored in the database to calculate the best location for a new pizza factory using as much of the host machine's resources as possible. Future implementations of this function in the future can account for road traffic during operating hours for a more accurate result. Visit the following url for mathematical assistance: https://stackoverflow.com/questions/6671183/calculate-the-center-point-of-multiple-latitude-longitude-coordinate-pairs

get_closest_factory(origin)

Calculates the shortest distance between the current location (origin) and a factory. Visit the following URL for mathematical assistance: https://www.mathwarehouse.com/algebra/distance_formula/index.php

get_config()

Reads the configuration from file

set_config(mongo_address, mongo_database, mongo_username, mongo_password, mongo_port, order_collection, factories_collection, latitude_field, longitude_field)

Writes the configuration to file